docs: update docs for task and some physical agents
